Resource based authorization completed

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesCommandHandler.cs
@@ -1,15 +1,18 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
 
 public class DeleteDishesCommandHandler(ILogger<DeleteDishesCommandHandler> logger,    
     IRestaurantsRepository resRepo,
-    IDishesRepository dishRepo) : IRequestHandler<DeleteDishesCommand>
+    IDishesRepository dishRepo,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishesCommand>
 {
     public async Task Handle(DeleteDishesCommand request, CancellationToken cancellationToken)
     {
@@ -20,6 +23,9 @@ public class DeleteDishesCommandHandler(ILogger<DeleteDishesCommandHandler> logg
         {
             throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
         }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidException();
 
         await dishRepo.Delete(restaurant.Dishes);
     }
